feat: add Sprint Calendar and List views, enhance SprintBoardContainer with task management features
- Implemented SprintCalendarView component for displaying tasks in a calendar format.
- Created SprintListView component for listing tasks with drag-and-drop functionality.
- Enhanced SprintBoardContainer to support multiple views (board, list, calendar) and task filtering.
- Added SprintSearch component for searching tasks and displaying sprint statistics.
- Introduced SprintTitle component for sprint management actions and status updates.
- Updated notifications service to fetch project-specific notifications.
- Modified task types to ensure priority field consistency.
- Added utility function to get column colors based on task status.
- Improved middleware to manage user access to routes based on authentication status.#15
