+++ src/lib/services/notifications.service.ts
@@ -9,5 +9,16 @@ export const notificationApi = {
   getUserNotifications: async () => {
     const response = await axiosInstance.get(`/notifications/`)
     return response.data
+  },
+
+  /**
+   * Get notifications for a specific project
+   * @param projectId The ID of the project
+   * @returns A list of notifications for the project
+   */
+  
+  getProjectNotifications: async (projectId: string) => {
+    const response = await axiosInstance.get(`/notifications/project/${projectId}`)
+    return response.data
   }
 }
